pyproject.toml
[tool.poetry]
name = "Flandre"
version = "0.4.0"
description = "基于nonebot2的机器人"
authors = ["KoishiChan <68314080+KoishiStudio@users.noreply.github.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "2.0.0b2"
nonebot-adapter-onebot = "2.0.0.b1"
aiohttp = "~=3.7.4.post0"
pixivpy-async = "~=1.2.14"
aiohttp-socks = "~=0.7.1"
bilibili-api = "9.1.0"
nonebot-plugin-help = "~=0.2.1"
"ruamel.yaml" = "~=0.17.21"
defusedxml = "~=0.7.1"
qrcode = "~=7.3.1"
Pillow = "~=9.0.1"
wcwidth = "~=0.2.5"
aiofile = "~=3.7.4"
nonebot-plugin-mediawiki = ">=0.1.4,<0.2.0"
nonebot-plugin-giyf = ">=0.1.1,<0.2.0"
jieba = "^0.42.1"
py7zr = "^0.18.3"
gTTS = "^2.2.4"

[tool.poetry.dev-dependencies]
nb-cli = "~=0.6.6"

[nonebot.plugins]
plugins = ["nonebot_plugin_help", "nonebot_plugin_mediawiki", "nonebot_plugin_giyf"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
